# writable/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Codeigniter Api",
        "description": "Example swagger openApi",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http:\/\/localhost:8080"
        }
    ],
    "paths": {
        "\/api\/users": {
            "get": {
                "operationId": "bbd2f6548aa41f9fdf76806379daa28f",
                "responses": {
                    "200": {
                        "description": "list users",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/User"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Not allowed"
                    }
                }
            }
        },
        "\/api\/test": {
            "get": {
                "operationId": "ad5635d8d731ab1ec38733221f7f7d03",
                "responses": {
                    "200": {
                        "description": "AOK"
                    },
                    "401": {
                        "description": "Not allowed"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "User": {
                "type": "array",
                "items": {
                    "properties": {
                        "id": {
                            "description": "Identifiant unique de l'utilisateur",
                            "type": "integer"
                        },
                        "name": {
                            "description": "Nom de l'utilisateur",
                            "type": "string"
                        },
                        "email": {
                            "description": "Email de l'utilisateur",
                            "type": "string"
                        },
                        "created_at": {
                            "description": "Date de cr\u00e9ation de l'utilisateur",
                            "type": "string",
                            "format": "date-time"
                        }
                    },
                    "type": "object"
                }
            }
        },
        "securitySchemes": {
            "securityScheme": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}